+++ fuzz/fuzz_targets/ethereum_types.rs
@@ -2,7 +2,7 @@
 
 use libfuzzer_sys::fuzz_target;
 
-use const_varint::{consume_varint, Varint};
+use varing::{consume_varint, Varint};
 use ethereum_types::{U128, U256, U512, U64};
 
 macro_rules! fuzzy {