+++ fuzz/fuzz_targets/primitive_types.rs
@@ -2,7 +2,7 @@
 
 use libfuzzer_sys::fuzz_target;
 
-use const_varint::{consume_varint, Varint};
+use varing::{consume_varint, Varint};
 use primitive_types::{U128, U256, U512};
 
 macro_rules! fuzzy {