+++ fuzz/fuzz_targets/ruint.rs
@@ -2,7 +2,7 @@
 
 use libfuzzer_sys::fuzz_target;
 
-use const_varint::{consume_varint, Varint};
+use varing::{consume_varint, Varint};
 use ruint::aliases::{
   U0, U1, U1024, U128, U16, U2048, U256, U32, U320, U384, U4096, U448, U512, U64, U768,
 };