+++ fuzz/fuzz_targets/time.rs
@@ -2,7 +2,7 @@
 
 use ::chrono::{Datelike, Timelike};
 use ::time::{Date, Duration, Month, PrimitiveDateTime, Time, UtcDateTime};
-use const_varint::*;
+use varing::*;
 use libfuzzer_sys::fuzz_target;
 
 macro_rules! fuzzy {