+++ src/num_complex.rs
@@ -1,0 +1,4 @@
+#[cfg(feature = "num-complex_0_4")]
+mod v04;
+#[cfg(feature = "num-complex_0_4")]
+pub use v04::*;