+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 FEATURES
 
 - Add `chrono` and `time` support
+- Add `num-rational` and `num-complex` support
+- Rename the crate from `const-varint` to `varing`
 
 ## 0.4.0 (Mar 22nd, 2025)